Use a data structure to represent the installed packages.
The problem I'm trying to solve is this:

1. Homebrew doesn't really expect a case-sensitive type system, so it
requires package directories to be in all-lowercase.

2. Cabal and Hackage *are* case-sensitive.

Because of this, some packages could be installed, but would spit out an
error message whenever I tried to see if it needed updating.

The solution I'm using is to include a JSON data file in each package
directory that contains the case-sensitive name, the version, and other
information. This commit implements that.